pyproject.toml
[project]
name = "pyplatformer"
version = "0.1.0"
description = ""
authors = [
    {name = "y"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp (==3.12.12)",
    "pygame-ce (==2.5.5)",
    "click (==8.2.1)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.13"
mypy = "^1.16.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py311"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.ruff.lint]
select = [
    "I",     # isort
    "N",     # pep8 naming
    "E",     # pycode style error
    "W",     # pycode style warning
    "F",     # pyflake
    "PL",    # pylint
    "C90",   # mccabe
    "PERF",  # perflint
]
